requirements.txt
asgiref==3.8.1
asttokens==2.4.1
bcrypt==4.2.0
cffi==1.17.0
cryptography==43.0.1
decorator==5.1.1
Deprecated==1.2.14
dj-database-url==2.2.0
Django==5.1
django-extensions==3.2.3
django-hosts==6.0
djangorestframework==3.15.2
executing==2.1.0
fabric==3.2.2
invoke==2.2.0
ipython==8.27.0
jedi==0.19.1
matplotlib-inline==0.1.7
paramiko==3.4.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
PyNaCl==1.5.0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
wrapt==1.16.0
